Copula Parameter of a Bivariate Copula for a Given Value of the Calibration Function

Computes the (first) copula parameter of a bivariate copula for a given value of the calibration function (eta).

BiCopEta2Par(family, eta)

Arguments

family
A copula family: 1 Gaussian, 2 Student t, 301 Double Clayton type I (standard and rotated 90 degrees), 302 Double Clayton type II (standard and rotated 270 degrees), 303 Double Clayton type III (survival and rotated 90 degrees), 304 Double Clayton type IV (survival and rotated 270 degrees), 401 Double Gumbel type I (standard and rotated 90 degrees), 402 Double Gumbel type II (standard and rotated 270 degrees), 403 Double Gumbel type III (survival and rotated 90 degrees), 404 Double Gumbel type IV (survival and rotated 270 degrees).

eta
The calibration function.

Value

The value of the first copula parameter, depending on the copula parameter and family as:

1 Gaussian, f(x) = tanh(x/2)

2 Student t, f(x) = tanh(x/2)

301 Double Clayton type I (standard and rotated 90 degrees), f(x) = x

302 Double Clayton type II (standard and rotated 270 degrees), f(x) = x

303 Double Clayton type III (survival and rotated 90 degrees), f(x) = x

304 Double Clayton type IV (survival and rotated 270 degrees), f(x) = x

401 Double Gumbel type I (standard and rotated 90 degrees), f(x) = x*(1+abs(x))/abs(x)

402 Double Gumbel type II (standard and rotated 270 degrees), f(x) = x*(1+abs(x))/abs(x)

403 Double Gumbel type III (survival and rotated 90 degrees), f(x) = x*(1+abs(x))/abs(x)

404 Double Gumbel type IV (survival and rotated 270 degrees) f(x) = x*(1+abs(x))/abs(x).

See also

BiCopEta2Par or BiCopPar2Tau and BiCopTau2Par from VineCopula.