Construction of a gamBiCop Class Object

Constructs an object of the class gamBiCop.

gamBiCop(family, model, par2 = 0, tau = TRUE)

Arguments

family
A copula family: 1 Gaussian, 2 Student t, 5 Frank, 301 Double Clayton type I (standard and rotated 90 degrees), 302 Double Clayton type II (standard and rotated 270 degrees), 303 Double Clayton type III (survival and rotated 90 degrees), 304 Double Clayton type IV (survival and rotated 270 degrees), 401 Double Gumbel type I (standard and rotated 90 degrees), 402 Double Gumbel type II (standard and rotated 270 degrees), 403 Double Gumbel type III (survival and rotated 90 degrees), 404 Double Gumbel type IV (survival and rotated 270 degrees).

model
A gamObject as return by the gam function from the mgcv package.

par2
Second parameter for the Studen t-copula.

tau
FALSE for a calibration fonction specified for the Copula parameter or TRUE (default) for a calibration function specified for Kendall's tau.

Value

An object of the class gamBiCop.

See also

gamBiCop, gamBiCopFit, gamBiCopPredict and gamBiCopSimulate.